move model methods to resource which uses them
