+++ internal/service/federatedsettingsorgconfig/data_source_federated_settings_connected_orgs.go
@@ -8,7 +8,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/federatedsettingsidentityprovider"
 	matlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
@@ -172,7 +171,7 @@ func flattenFederatedSettingsOrganizationConfigs(federatedSettingsConnectedOrgan
 					"identity_provider_id":       federatedSettingsConnectedOrganizations.Results[i].IdentityProviderID,
 					"org_id":                     federatedSettingsConnectedOrganizations.Results[i].OrgID,
 					"post_auth_role_grants":      federatedSettingsConnectedOrganizations.Results[i].PostAuthRoleGrants,
-					"role_mappings":              federatedsettingsidentityprovider.FlattenRoleMappings(federatedSettingsConnectedOrganizations.Results[i].RoleMappings),
+					"role_mappings":              FlattenRoleMappings(federatedSettingsConnectedOrganizations.Results[i].RoleMappings),
 					"user_conflicts":             nil,
 				}
 			} else {
@@ -182,8 +181,8 @@ func flattenFederatedSettingsOrganizationConfigs(federatedSettingsConnectedOrgan
 					"identity_provider_id":       federatedSettingsConnectedOrganizations.Results[i].IdentityProviderID,
 					"org_id":                     federatedSettingsConnectedOrganizations.Results[i].OrgID,
 					"post_auth_role_grants":      federatedSettingsConnectedOrganizations.Results[i].PostAuthRoleGrants,
-					"role_mappings":              federatedsettingsidentityprovider.FlattenRoleMappings(federatedSettingsConnectedOrganizations.Results[i].RoleMappings),
-					"user_conflicts":             federatedsettingsidentityprovider.FlattenUserConflicts(*federatedSettingsConnectedOrganizations.Results[i].UserConflicts),
+					"role_mappings":              FlattenRoleMappings(federatedSettingsConnectedOrganizations.Results[i].RoleMappings),
+					"user_conflicts":             FlattenUserConflicts(*federatedSettingsConnectedOrganizations.Results[i].UserConflicts),
 				}
 			}
 		}
