+++ internal/service/federatedsettingsorgrolemapping/data_source_federated_settings_org_role_mapping.go
@@ -8,7 +8,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/federatedsettingsidentityprovider"
 )
 
 func DataSource() *schema.Resource {
@@ -90,7 +89,7 @@ func dataSourceMongoDBAtlasFederatedSettingsOrganizationRoleMappingRead(ctx cont
 		return diag.FromErr(fmt.Errorf("error setting `result` for federatedSettings Role Mapping: %s", err))
 	}
 
-	if err := d.Set("role_assignments", federatedsettingsidentityprovider.FlattenRoleAssignments(federatedSettingsOrganizationRoleMapping.RoleAssignments)); err != nil {
+	if err := d.Set("role_assignments", FlattenRoleAssignments(federatedSettingsOrganizationRoleMapping.RoleAssignments)); err != nil {
 		return diag.FromErr(fmt.Errorf("error setting `result` for federatedSettings Role Mapping: %s", err))
 	}
